Moved util tests to unit test folder
