+++ src/tests/unit/utils/text-utils.test.ts
@@ -1,6 +1,26 @@
 import { expect, describe, it } from 'vitest';
 
-import { pascalCaseToText, textToKebabCase, printable, mergePropertiesPath } from '../text-utils';
+import { getPathArray, pascalCaseToText, textToKebabCase, printable, mergePropertiesPath } from '#/app/utils/text-utils';
+
+describe('getPathArray', () => {
+  it('should work with path literals', () => {
+    const path = 'position.y';
+    const result = getPathArray(path);
+    expect(result).toEqual(['position', 'y']);
+  });
+
+  it('should work with regex paths', () => {
+    const path = '/.*/.color';
+    const result = getPathArray(path);
+    expect(result).toEqual(['/.*/', 'color']);
+  });
+
+  it('should respect the order of the path', () => {
+    const path = 'position./.*/';
+    const result = getPathArray(path);
+    expect(result).toEqual(['position', '/.*/']);
+  });
+});
 
 describe('pascalCaseToText', () => {
   it('should convert pascal case to text', () => {
@@ -39,4 +59,4 @@ describe('mergePropertiesPath', () => {
   it('should return only key when path is undefined', () => {
     expect(mergePropertiesPath('key', undefined)).toBe('key');
   });
-}); 
+});
